Repair native offline portability without weakening research contracts
Continue PR #114 on the same canonical integration line. Preserve all original CSV/ledger/model/installed-consumer assertions. Accept strictly parsed finite subnormal CSV values across standard libraries; keep malformed/overflow/permanent-failure behavior. On narrow-long-double ABIs retain settlement precision guards with an internal bounded 64-significand accumulator in the existing ledger. Add exact arithmetic, subnormal, rollback and installed-consumer regressions; retain native extended-long-double arithmetic on existing ABIs. Use an exclusive Windows user-temp delete-on-close output spool and explicit bounded fixture conversions. No second ledger/OMS/transport, production install change, broker path or legacy ABI claim.

Complete tree 50590a3fabdb98fcdfb4d25053e7f35278f05bd2 verified against the local ten-file repair. Digest-bound object authoring run 35763138481 uploaded only checked blobs/tree, never ran candidate code or wrote refs. A previous inert transport transcription failed before writes and was corrected with the same required patch/tree digests. This delivery removes the temporary authoring workflow; normal exact-head CI remains required. Real remote parent retained; synthetic local validation ancestry is not imported.

+++ tests/research/market_data_tests.cpp
@@ -699,7 +699,7 @@ void MergedCsvOracle() {
             std::int64_t timestamp = static_cast<std::int64_t>((source * 7 + trial) % 4);
             for (std::size_t row = 0; row < (trial + source * 3) % 14; ++row) {
                 timestamp += static_cast<std::int64_t>((row * 3 + source + trial) % 4);
-                Tick tick = T(timestamp, row + 1, 100 + source + row,
+                Tick tick = T(timestamp, row + 1, static_cast<double>(100 + source + row),
                               static_cast<std::int64_t>((row + trial) % 5));
                 tick.instrument = "S" + std::to_string(source);
                 ticks.push_back(tick); expected.push_back({tick, source});
